Reference

Types

SupplyChainOptimization.SupplyChain — Type

The supply chain.

SupplyChainOptimization.Location — Type

The geographical location of a node of the supply chain. The location is defined by its latitude and longitude.

SupplyChainOptimization.Node — Type

A node of the supply chain.

SupplyChainOptimization.Product — Type

A product.

SupplyChainOptimization.Customer — Type

A customer.

SupplyChainOptimization.Lane — Type

A transportation lane between two nodes of the supply chain.

SupplyChainOptimization.Plant — Type

A plant.

SupplyChainOptimization.Storage — Type

A storage location.

SupplyChainOptimization.Supplier — Type

A supplier.

SupplyChainOptimization.Demand — Type

The demand a customer has for a product.

Modeling

SupplyChainOptimization.add_plant! — Function

add_plant!(supply_chain, plant)

Adds a plant to the supply chain.

SupplyChainOptimization.add_supplier! — Function

add_supplier!(supply_chain, supplier)

Adds a supplier to the supply chain.

SupplyChainOptimization.add_storage! — Function

add_storage!(supply_chain, storage)

Adds a storage location to the supply chain.

SupplyChainOptimization.add_customer! — Function

add_customer!(supply_chain, customer)

Adds a customer to the supply chain.

SupplyChainOptimization.add_product! — Function

add_product!(customer::Customer, product::Product; demand::Array{Float64, 1})

Indicates that a customer needs a product.

The keyword arguments are:

demand: the amount of unit of product needed for each time period.

add_product!(supplier::Supplier, product::Product; unit_cost::Float64, maximum_throughput::Float64)

Indicates that a supplier can provide a product.

The keyword arguments are:

unit_cost: the cost per unit of the product from this supplier.

maximum_throughput: the maximum number of units that can be provided in each time period.

add_product!(plant::Plant, product::Product; bill_of_material::Dict{Product, Float64}, unit_cost, maximum_throughput)

Indicates that a plant can produce a product.

The keyword arguments are:

bill_of_material: the amount of other product needed to produce one unit of the product. This dictionary can be empty if there are no other products needed.

unit_cost: the cost of producing one unit of product.

maximum_throughput: the maximum amount of product that can be produced in a time period.

time: the production lead time.

add_product!(supply_chain, product)

Adds a product to the supply chain.

SupplyChainOptimization.add_demand! — Function

add_demand!(supply_chain, customer, product; demand::Array{Float64, 1}, service_level=1.0)

Adds customer demand for a product. The demand is specified for each time period.

The keyword arguments are:

demand: the amount of product demanded for each time period.

service_level: indicates how many lost sales are allowed as a ratio of demand. No demand can be lost if the service level is 1.0 and all demand can be lost if the service level is 0.0.

add_demand!(supply_chain, demand)

Adds demand to the supply chain.

SupplyChainOptimization.add_lane! — Function

add_lane!(supply_chain, lane)

Adds a transportation lane to the supply chain.

SupplyChainOptimization.haversine — Function

Computes the great circle distance between two locations. The distance is expressed in meter.

Optimization

SupplyChainOptimization.optimize_network! — Function

optimize_network!(supply_chain::SupplyChain, optimizer=HiGHS.Optimizer)

Optimizes the supply chain.

Querying Results

SupplyChainOptimization.get_total_costs — Function

get_total_costs(supply_chain::SupplyChain)

Gets the total costs of operating the supply chain.

SupplyChainOptimization.get_total_fixed_costs — Function

get_total_fixed_costs(supply_chain::SupplyChain)

Gets the total fixed costs of operating the supply chain.

SupplyChainOptimization.get_total_transportation_costs — Function

get_total_transportation_costs(supply_chain::SupplyChain)

Gets the total transportation costs of operating the supply chain.

SupplyChainOptimization.get_inventory_at_start — Function

Gets the inventory of a product stored at the start of a period.

SupplyChainOptimization.get_inventory_at_end — Function

Gets the inventory of a product stored at the end of a period.

SupplyChainOptimization.get_production — Function

get_production(supply_chain::SupplyChain, plant::Plant, product::Product, period=1)

Gets the amount of a given product produced at a given plant during a given period.

SupplyChainOptimization.get_receipts — Function

get_receipts(supply_chain::SupplyChain, storage::Storage, product::Product, period=1)

Gets the amount of a given product received at a given storage location at a given period.

SupplyChainOptimization.get_shipments — Function

get_shipments(supply_chain::SupplyChain, storage::Storage, product::Product, period=1)

Gets the amount of a given product sent from a given storage location at a given period.

get_shipments(supply_chain::SupplyChain, plant::Plant, product::Product, period=1)

Gets the amount of a given product sent from a given plant at a given period.

get_shipments(supply_chain::SupplyChain, supplier::Supplier, product::Product, period=1)

Gets the amount of a given product shipped from a given supplier at a given period.

get_shipments(supply_chain::SupplyChain, lane::Lane, product::Product, period=1)

Gets the amount of a given product sent on a lane at a given period.

SupplyChainOptimization.is_opened — Function

is_opened(supply_chain::SupplyChain, storage::Storage, period=1)

Gets whether a given storage location is opened during a given period.

is_opened(supply_chain::SupplyChain, plant::Plant, period=1)

Gets whether a given plant is opened during a given period.

SupplyChainOptimization.is_opening — Function

Gets whether a given storage location is opening during a given period.

SupplyChainOptimization.is_closing — Function

Gets whether a given storage location is closing during a given period.

Visualization

SupplyChainOptimization.plot_network — Function

plot_network(supply_chain, period=1; geography="usa", showlegend=true)

Plots the nodes of the supply chain on a map.

The geography must be one of: "world" | "usa" | "europe" | "asia" | "africa" | "north america" | "south america".

SupplyChainOptimization.plot_flows — Function

plot_flows(supply_chain, period=1; geography="usa", showlegend=true)

Plots the flows of products in the supply chain.

SupplyChainOptimization.plot_costs — Function

plot_costs(supply_chain)

Plots the costs of operating the supply chain.

SupplyChainOptimization.plot_inventory — Function

plot_inventory(supply_chain, storage, product)

Plots the amount of inventory of a product on-hand at a storage location at the beginning of each period.